Revert "fix(scraper): move service_controller out of services/railway/ package (#212)"
This reverts commit 0a055579becc46e18d9fbc0d058aee7ec4cacf2f.
